+++ include/seahorn/Expr/ExprOpCompare.hh
@@ -5,6 +5,7 @@
 #include "seahorn/Expr/ExprOpBool.hh"
 #include "seahorn/Expr/ExprOpCore.hh"
 #include "seahorn/Expr/TypeChecker.hh"
+#include "seahorn/Expr/TypeCheckerUtils.hh"
 
 namespace expr {
 
@@ -14,11 +15,15 @@ namespace typeCheck {
 namespace compareType {
 
 struct Equality {
+  // Return type: BOOL_TY
+  // Possible types of children: anything excepty for ERROR_TY
   static inline Expr inferType(Expr exp, TypeChecker &tc) {
     return typeCheck::binary<BOOL_TY, ANY_TY>(exp, tc);
   }
 };
 struct Inequality {
+  // Return type: BOOL_TY
+  // Possible types of children: INT_TY, REAL_TY, UNINT_TY
   static inline Expr inferType(Expr exp, TypeChecker &tc) {
     return typeCheck::binary<BOOL_TY, INT_TY, REAL_TY, UNINT_TY>(exp, tc);
   }
